% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{column.delete}
\alias{column.delete}
\title{Deletes the column.}
\usage{
column.delete(tableId, columnId)
}
\arguments{
\item{tableId}{Table from which the column is being deleted}

\item{columnId}{Name or identifier for the column being deleted}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

