% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{column.get}
\alias{column.get}
\title{Retrieves a specific column by its id.}
\usage{
column.get(tableId, columnId)
}
\arguments{
\item{tableId}{Table to which the column belongs}

\item{columnId}{Name or identifier for the column that is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

