% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Column}
\alias{Column}
\title{Column Object}
\usage{
Column(Column.baseColumn = NULL, baseColumn = NULL, columnId = NULL,
  description = NULL, graph_predicate = NULL, name = NULL, type = NULL)
}
\arguments{
\item{Column.baseColumn}{The \link{Column.baseColumn} object or list of objects}

\item{baseColumn}{Optional identifier of the base column}

\item{columnId}{Identifier for the column}

\item{description}{Optional column description}

\item{graph_predicate}{Optional column predicate}

\item{name}{Required name of the column}

\item{type}{Required type of the column}
}
\value{
Column object
}
\description{
Column Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the id, name and type of a column in a table.
}
\seealso{
Other Column functions: \code{\link{Column.baseColumn}},
  \code{\link{column.insert}}, \code{\link{column.patch}},
  \code{\link{column.update}}
}
