% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{ColumnList}
\alias{ColumnList}
\title{ColumnList Object}
\usage{
ColumnList(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{List of all requested columns}

\item{nextPageToken}{Token used to access the next page of this result}

\item{totalItems}{Total number of columns for the table}
}
\value{
ColumnList object
}
\description{
ColumnList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a list of columns in a table.
}

