% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\docType{package}
\name{fusiontables_googleAuthR}
\alias{fusiontables_googleAuthR}
\alias{fusiontables_googleAuthR-package}
\title{Fusion Tables API
API for working with Fusion Tables data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:51:13
filename: /Users/mark/dev/R/autoGoogleAPI/googlefusiontablesv1.auto/R/fusiontables_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}
}
