% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Import}
\alias{Import}
\title{Import Object}
\usage{
Import(numRowsReceived = NULL)
}
\arguments{
\item{numRowsReceived}{The number of rows received from the import request}
}
\value{
Import object
}
\description{
Import Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an import request.
}

