% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{PointStyle}
\alias{PointStyle}
\title{PointStyle Object}
\usage{
PointStyle(iconName = NULL, iconStyler = NULL)
}
\arguments{
\item{iconName}{Name of the icon}

\item{iconStyler}{Column or a bucket value from which the icon name is to be determined}
}
\value{
PointStyle object
}
\description{
PointStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a PointStyle within a StyleSetting
}

