% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{query.sqlGet}
\alias{query.sqlGet}
\title{Executes an SQL SELECT/SHOW/DESCRIBE statement.}
\usage{
query.sqlGet(sql, hdrs = NULL, typed = NULL)
}
\arguments{
\item{sql}{An SQL SELECT/SHOW/DESCRIBE statement}

\item{hdrs}{Should column names be included (in the first row)?}

\item{typed}{Should typed values be returned in the (JSON) response -- numbers for numeric values and parsed geometries for KML values? Default is true}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}
