% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{StyleFunction.gradient}
\alias{StyleFunction.gradient}
\title{StyleFunction.gradient Object}
\usage{
StyleFunction.gradient(StyleFunction.gradient.colors = NULL, colors = NULL,
  max = NULL, min = NULL)
}
\arguments{
\item{StyleFunction.gradient.colors}{The \link{StyleFunction.gradient.colors} object or list of objects}

\item{colors}{Array with two or more colors}

\item{max}{Higher-end of the interpolation range: rows with this value will be assigned to colors[n-1]}

\item{min}{Lower-end of the interpolation range: rows with this value will be assigned to colors[0]}
}
\value{
StyleFunction.gradient object
}
\description{
StyleFunction.gradient Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Gradient function that interpolates a range of colors based on column value.
}
\seealso{
Other StyleFunction functions: \code{\link{StyleFunction.gradient.colors}},
  \code{\link{StyleFunction}}
}

