% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{StyleFunction}
\alias{StyleFunction}
\title{StyleFunction Object}
\usage{
StyleFunction(StyleFunction.gradient = NULL,
  StyleFunction.gradient.colors = NULL, buckets = NULL, columnName = NULL,
  gradient = NULL)
}
\arguments{
\item{StyleFunction.gradient}{The \link{StyleFunction.gradient} object or list of objects}

\item{StyleFunction.gradient.colors}{The \link{StyleFunction.gradient.colors} object or list of objects}

\item{buckets}{Bucket function that assigns a style based on the range a column value falls into}

\item{columnName}{Name of the column whose value is used in the style}

\item{gradient}{Gradient function that interpolates a range of colors based on column value}
}
\value{
StyleFunction object
}
\description{
StyleFunction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a StyleFunction within a StyleSetting
}
\seealso{
Other StyleFunction functions: \code{\link{StyleFunction.gradient.colors}},
  \code{\link{StyleFunction.gradient}}
}

