% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{table.copy}
\alias{table.copy}
\title{Copies a table.}
\usage{
table.copy(tableId, copyPresentation = NULL)
}
\arguments{
\item{tableId}{ID of the table that is being copied}

\item{copyPresentation}{Whether to also copy tabs, styles, and templates}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}
