% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Table}
\alias{Table}
\title{Table Object}
\usage{
Table(attribution = NULL, attributionLink = NULL, baseTableIds = NULL,
  columns = NULL, description = NULL, isExportable = NULL, name = NULL,
  sql = NULL, tableId = NULL)
}
\arguments{
\item{attribution}{Optional attribution assigned to the table}

\item{attributionLink}{Optional link for attribution}

\item{baseTableIds}{Optional base table identifier if this table is a view or merged table}

\item{columns}{Columns in the table}

\item{description}{Optional description assigned to the table}

\item{isExportable}{Variable for whether table is exportable}

\item{name}{Name assigned to a table}

\item{sql}{Optional sql that encodes the table definition for derived tables}

\item{tableId}{Encrypted unique alphanumeric identifier for the table}
}
\value{
Table object
}
\description{
Table Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a table. Specifies the name, whether it is exportable, description, attribution, and attribution link.
}
\seealso{
Other Table functions: \code{\link{table.insert}},
  \code{\link{table.patch}}, \code{\link{table.update}}
}
