% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{table.update}
\alias{table.update}
\title{Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated.}
\usage{
table.update(Table, tableId, replaceViewDefinition = NULL)
}
\arguments{
\item{Table}{The \link{Table} object to pass to this method}

\item{tableId}{ID of the table that is being updated}

\item{replaceViewDefinition}{Should the view definition also be updated? The specified view definition replaces the existing one}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}

Other Table functions: \code{\link{Table}},
  \code{\link{table.insert}}, \code{\link{table.patch}}
}
