% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Task}
\alias{Task}
\title{Task Object}
\usage{
Task(progress = NULL, started = NULL, taskId = NULL, type = NULL)
}
\arguments{
\item{progress}{An indication of task progress}

\item{started}{false while the table is busy with some other task}

\item{taskId}{Identifier for the task}

\item{type}{Type of background task}
}
\value{
Task object
}
\description{
Task Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the identifier, name, and type of a task in a table.
}

