% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{TaskList}
\alias{TaskList}
\title{TaskList Object}
\usage{
TaskList(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{List of all requested tasks}

\item{nextPageToken}{Token used to access the next page of this result}

\item{totalItems}{Total number of tasks for the table}
}
\value{
TaskList object
}
\description{
TaskList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a list of tasks for a table.
}
