% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{template.update}
\alias{template.update}
\title{Updates an existing template}
\usage{
template.update(Template, tableId, templateId)
}
\arguments{
\item{Template}{The \link{Template} object to pass to this method}

\item{tableId}{Table to which the updated template belongs}

\item{templateId}{Identifier for the template that is being updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}

Other Template functions: \code{\link{Template}},
  \code{\link{template.insert}},
  \code{\link{template.patch}}
}

