% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{TemplateList}
\alias{TemplateList}
\title{TemplateList Object}
\usage{
TemplateList(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{List of all requested templates}

\item{nextPageToken}{Token used to access the next page of this result}

\item{totalItems}{Total number of templates for the table}
}
\value{
TemplateList object
}
\description{
TemplateList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a list of templates for a given table.
}

